refactor: renomeia pasta backend para markupp e backendUrl para serverUrl
O servidor Go é o produto do monorepo; não há frontend no repo. O nome
backend induzia leitura errada da arquitetura.

- backend/ vira markupp/, module path e imports Go ajustados
- docker-compose e CI apontam para a nova pasta
- backendUrl vira serverUrl no plugin Obsidian (settings, cliente, mensagens)
- ADR-0001 renomeada para linguagem-servidor; ADR-0004 e docs trocam
  "backend" conceitual por "servidor"
- openapi: title "Markupp Backend API" vira "Markupp Server API"

Closes #70
