+++ markupp/internal/notes/service_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ifsc-ES2/projeto-markupp/backend/internal/notes"
+	"github.com/ifsc-ES2/projeto-markupp/markupp/internal/notes"
 )
 
 const testMaxContentSize = 100
