+++ markupp/openapi.yaml
@@ -1,6 +1,6 @@
 openapi: 3.0.3
 info:
-  title: Markupp Backend API
+  title: Markupp Server API
   description: API REST para gerenciamento de notas do projeto Markupp.
   version: 0.1.0
 servers:
