+++ obsidian-plugin/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "obsidian-markupp-plugin",
 	"version": "0.1.0",
-	"description": "Plugin Obsidian do projeto Markupp — envia notas para o backend self-hosted.",
+	"description": "Plugin Obsidian do projeto Markupp: envia notas para o servidor Markupp self-hosted.",
 	"main": "main.js",
 	"type": "module",
 	"scripts": {
