+++ obsidian-plugin/src/commands/upload-active-note.ts
@@ -17,14 +17,14 @@ export async function uploadActiveNote(
 
 	try {
 		const content = await plugin.app.vault.read(file);
-		await createNote(settings.backendUrl, path, content);
+		await createNote(settings.serverUrl, path, content);
 		new Notice(`Nota enviada: ${path}`);
 	} catch (err) {
-		new Notice(buildErrorMessage(err, settings.backendUrl));
+		new Notice(buildErrorMessage(err, settings.serverUrl));
 	}
 }
 
-function buildErrorMessage(err: unknown, backendUrl: string): string {
+function buildErrorMessage(err: unknown, serverUrl: string): string {
 	if (err instanceof MarkuppApiError) {
 		switch (err.code) {
 			case "invalid_path":
@@ -37,5 +37,5 @@ function buildErrorMessage(err: unknown, backendUrl: string): string {
 				return `Erro do servidor (${err.status}): ${err.message}`;
 		}
 	}
-	return `Não foi possível conectar a ${backendUrl}. Verifique se o backend está rodando.`;
+	return `Não foi possível conectar a ${serverUrl}. Verifique se o servidor está rodando.`;
 }
